refactor(ocpp): extract setup-dialog security domain into security.rs
Stage 1 of #89. Moves SecurityLevel/SkipVerifyChoice/SecurityInputs, from_config/
build_config, and validate_security (with their from_config/build_config tests) out of
the dialog file. setup_dialog.rs 1722->1327.
